pyproject.toml
[project]
name = "quarterly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "llama-index-core>=0.14.16",
    "llama-index-embeddings-ollama>=0.8.6",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-vector-stores-chroma>=0.5.5",
    "uvicorn>=0.41.0",
]
 
[project.scripts]
quarterly = "quarterly.main:main"


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
